requirements_runpod.txt
# Requirements for RunPod vLLM Docker deployment
# Core dependencies
runpod>=1.6.5
vllm>=0.6.0

# Image processing
pillow>=10.1.0
pypdf2>=3.0.1
filetype>=1.2.0

# Chandra OCR (includes torch, transformers, etc.)
--extra-index-url https://download.pytorch.org/whl/cu121
chandra-ocr>=0.1.8

# Additional utilities
pydantic>=2.0.0
